add tmux plugin yank, add zsh setting

+++ install_to_unix.sh
@@ -1,22 +1,47 @@
 #!/bin/bash
 
-ln -s ~/dotfiles-193/vim/.vimrc ~/.vimrc
-ln -s ~/dotfiles-193/vim/.gvimrc ~/.gvimrc
-ln -s ~/dotfiles-193/.tmux.conf ~/.tmux.conf
-ln -s ~/dotfiles-193/.bash_aliases ~/.bash_aliases
-echo "source ~/.bashrc" >> ~/.bash_profile
-echo "source ~/.bash_aliases" >> ~/.bashrc
+if [ ! -e ~/.vimrc ] ; then
+  ln -s ~/dotfiles-193/vim/.vimrc ~/.vimrc
+fi
 
-# install dein
-curl https://raw.githubusercontent.com/Shougo/dein.vim/master/bin/installer.sh > ~/installer.sh
-source ~/installer.sh ~/.cache/dein
-rm ~/installer.sh
+if [ ! -e ~/.gvimrc ] ; then
+  ln -s ~/dotfiles-193/vim/.gvimrc ~/.gvimrc
+fi
 
-# install volta
-curl https://get.volta.sh | bash
+if [ ! -e ~/.tmux.conf ] ; then
+  ln -s ~/dotfiles-193/.tmux.conf ~/.tmux.conf
+fi
 
-# install tpm
-git clone https://github.com/tmux-plugins/tpm ~/.tmux/plugins/tpm
+if [ ! -e ~/.bash_aliases ] ; then
+  ln -s ~/dotfiles-193/.bash_aliases ~/.bash_aliases
+fi
+
+if [ $SHELL = "/bin/bash" ] ; then
+  if `! cat ~/.bash_profile | grep "source ~/.bash_aliases"` ; then
+    echo "source ~/.bash_aliases" >> ~/.bash_profile
+  fi
+elif [ $SHELL = "/bin/zsh" ] ; then
+  if `! cat ~/.zshrc | grep "source ~/.bash_aliases"` ; then
+    echo "source ~/.bash_aliases" >> ~/.zshrc
+  fi
+fi
+
+if [ ! -e ~/.cache/dein ] ; then
+  # install dein
+  curl https://raw.githubusercontent.com/Shougo/dein.vim/master/bin/installer.sh > ~/installer.sh
+  source ~/installer.sh ~/.cache/dein
+  rm ~/installer.sh
+fi
+
+if [ ! -e ~/.volta ] ; then
+  # install volta
+  curl https://get.volta.sh | bash
+fi
+
+if [ ! -e ~/.tmux/plugins/tpm ] ; then
+  # install tpm
+  git clone https://github.com/tmux-plugins/tpm ~/.tmux/plugins/tpm
+fi
 
 # font のインストール
 # git clone --depth 1 https://github.com/ryanoasis/nerd-fonts
